Update identification/pemethod.py guarantee that nkk is integer in armax

+++ pysid/identification/pemethod.py
@@ -198,7 +198,7 @@ def armax(na, nb, nc, nk, u, y):
         index = delete(index, i)
         # Inputs
         inps = concatenate((y[:, index], u, ehat), axis=1)
-        nkk = array(append([1, ]*len(na[i, index]), append(nk[i, :], 1)), ndmin=2)
+        nkk = array(append([1, ]*len(na[i, index]), append(nk[i, :], 1)), ndmin=2, dtype='int')
         A_, BAC = arx([na[i, i]], array(append(na[i, index] - 1, append(nb[i, :], nc[i]-1)), ndmin=2), nkk, inps, y[:, i:i+1])
         thetai = A_[0][0][1:]
         for k in range(len(nkk[0])):
